Remove confusing line of code that does nothing
I asked about this line in the SpongePowered Discord, and I was told by @Aaron1011 that it "will modify a fresh DataContainer which isn't connected to the skull".  He mentioned that the proper way is to use RepresentedPlayerData -- but the REPRESENTED_PLAYER key is already set to `user.getProfile()`, so even if the line did anything it would be redundant.